Split Active Module into Heal, Slow, melee and range subclasses
