+++ Assets/Scripts/Modules/MeleeAttackModule.cs
@@ -1,0 +1,30 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MeleeAttackModule : Module {
+
+    //basic constructor (give name)
+    //We can change stats as necessary for balance
+    public MeleeAttackModule()
+    {
+        ModuleName = ModuleName.shortRange;
+        HitPoints = 25;
+        Mass = 100;
+        action = new Action(ActionType.ShortAttack, 25, 1);
+    }
+
+    public override Action GetAction()
+    {
+        return action;
+    }
+    // Use this for initialization
+    void Start () {
+		
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		
+	}
+}
