+++ Assets/Scripts/Modules/MeleeAttackModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e09e57de80cb06a458735346779bc4f7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
