+++ Assets/Scripts/Modules/RangeAttackModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58a52289d80bc464f8571628abe35b5b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
