+++ Assets/Scripts/Modules/SlowModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e609ff7eb590a6b4290044f5d842c2c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
